cifr_agent_system/requirements.txt
google-cloud-aiplatform
google-cloud-dlp
google-cloud-pubsub
google-cloud-storage
google-cloud-language
google-genai # Added for direct Gemini API access
cryptography
pillow
python-dotenv # For local environment variable management
Flask[async]
Flask[async] # For the web UI
